Signatures update.
This signatures update also removes any references to the Google Malware
Dashboard from the signature files. According to a message displayed at the
webpage where that information is sourced from, the page in question has
been archived since April 8th, 2020 (so slightly over a month now), and
there doesn't seem to be any particular indication that the information
found there will continue to be updated in the future, so keeping it in the
signature files is kind of pointless, seeing as it would likely stale very
quickly.