BOBUAM for v3 initial commit.
Commits BOBUAM as it was at the time that it support had ended and it had been
taken offline (2021.81.478), functional for v2 and earlier, but predating
CIDRAM v3 and not yet functional for v3.

Committed for the benefit of providing unambiguous diffs to show what's
changed between that version and and its reinstatement for CIDRAM v3 to come.